refactor: break into some files
