+++ src/app/(app)/(layout)/wallet/send/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Send } from '@/views/wallet/Send';
+import { Send } from '@/views/wallet/send/Send';
 
 export default function Page() {
   return <Send />;
