+++ src/views/wallet/send/assets.ts
@@ -1,0 +1,3 @@
+export type Assets = 'Liquid Bitcoin' | 'Tether USD';
+
+export const assets: Assets[] = ['Liquid Bitcoin', 'Tether USD'];
